+++ release_notes.md
@@ -1,4 +1,6 @@
 ###In Development
+  - [#147](https://github.com/MehdiK/Humanizer/pull/147): Added Russian translation for ToWords
+
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.20.15...master)
 
 ###v1.20.15 - 2014-04-12